requirement.txt
numpy
flask_restx
opencv-python
Flask-Migrate==2.1.1
Flask-Script==2.0.6

Flask-Bcrypt==0.7.1
flask-restplus==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
SQLAlchemy==1.2.5
Werkzeug